profile restyling + pizazz, moving assets into profile and home folders
