+++ frontend/src/assets/home/cartoon-frame.png
